chore: ajustado logica e retirado promisse
