+++ src/index.js
@@ -8,6 +8,6 @@ app.use(cors());
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 
-require('./controllers/productController')(app);
+require('./controllers/product-controller')(app);
 
 app.listen(process.env.PORT || 8000);
